feat: adapt rules and report to new structure
