+++ workflow/envs/merge_json.yaml
@@ -1,0 +1,7 @@
+channels:
+  - conda-forge
+  - bioconda
+  - defaults
+dependencies:
+  - pip:
+    - cyvcf2==0.30.15
