+++ workflow/rules/cnv_html_report.smk
@@ -1,39 +1,35 @@
+def get_cnv_ratio_file(wildcards):
+    caller = wildcards.get("caller", "")
+    if caller == "cnvkit":
+        return "cnv_sv/cnvkit_batch/{sample}/{sample}_{type}.cnr".format(**wildcards)
+    elif caller == "gatk_cnv":
+        return "cnv_sv/gatk_cnv_denoise_read_counts/{sample}_{type}.clean.denoisedCR.tsv".format(**wildcards)
+    else:
+        raise NotImplementedError(f"not implemented for caller {caller}")
+
+
+def get_cnv_segment_file(wildcards):
+    caller = wildcards.get("caller", "")
+    if caller == "cnvkit":
+        return "cnv_sv/cnvkit_batch/{sample}/{sample}_{type}.cns".format(**wildcards)
+    elif caller == "gatk_cnv":
+        return "cnv_sv/gatk_cnv_model_segments/{sample}_{type}.clean.cr.seg".format(**wildcards)
+    else:
+        raise NotImplementedError(f"not implemented for caller {caller}")
+
+
 rule cnv_json:
     input:
-        amp_bed=config.get("annotate_cnv", {}).get("cnv_amp_genes", []),
-        cnr="cnv_sv/cnvkit_batch/{sample}/{sample}_{type}.cnr",
-        cns="cnv_sv/cnvkit_batch/{sample}/{sample}_{type}.cns",
-        gatk_ratios="cnv_sv/gatk_denoise_read_counts/{sample}_{type}.clean.denoisedCR.tsv",
-        gatk_segments="cnv_sv/gatk_model_segments/{sample}_{type}.clean.cr.seg",
-        germline_vcf="snv_indels/bcbio_variation_recall_ensemble/{sample}_{type}.ensembled.vep_annotated.filter.germline.vcf",
-        fai=config.get("reference").get("fai"),
-        loh_bed=config.get("annotate_cnv", {}).get("cnv_loh_genes", []),
-        svdb_vcfs=[
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_amp_genes.vcf.gz",
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_loh_genes.vcf.gz",
-        ],
-        svdb_tbis=[
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_amp_genes.vcf.gz.tbi",
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_loh_genes.vcf.gz.tbi",
-        ],
-        svdb_filtered_vcfs=[
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_amp_genes.filter.cnv_hard_filter_amp.vcf.gz",
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_loh_genes.filter.cnv_hard_filter_loh.vcf.gz",
-        ],
-        svdb_filtered_tbis=[
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_amp_genes.filter.cnv_hard_filter_amp.vcf.gz.tbi",
-            "cnv_sv/svdb_query/{sample}_{type}.svdb_query.annotate_cnv.cnv_loh_genes.filter.cnv_hard_filter_loh.vcf.gz.tbi",
-        ],
+        ratios=get_cnv_ratio_file,
+        segments=get_cnv_segment_file,
     output:
-        json=temp("cnv_sv/cnv_html_report/{sample}_{type}.cnv.json"),
-    params:
-        skip_chromosomes=config.get("reference", {}).get("skip_chrs", None),
+        json=temp("cnv_sv/cnv_html_report/{sample}_{type}.{caller}.json"),
     log:
-        "cnv_sv/cnv_html_report/{sample}_{type}.cnv.json.log",
+        "cnv_sv/cnv_html_report/{sample}_{type}.{caller}.cnv_json.log",
     benchmark:
         repeat(
-            "cnv_sv/cnv_html_report/{sample}_{type}.cnv.json.benchmark.tsv",
-            config.get("cnv_json", {}).get("benchmark_repeats", 1),
+            "cnv_sv/cnv_html_report/{sample}_{type}.{caller}.cnv_json.benchmark.tsv",
+            config.get("merge_json", {}).get("benchmark_repeats", 1),
         )
     threads: config.get("cnv_json", {}).get("threads", config["default_resources"]["threads"])
     resources:
@@ -47,14 +43,55 @@ rule cnv_json:
     conda:
         "../envs/cnv_json.yaml"
     message:
-        "{rule}: Create a JSON representation of the cnv results: {output.json}"
+        "{rule}: Create JSON representation of CNV results: {output.json}"
     script:
         "../scripts/cnv_json.py"
 
 
+rule merge_json:
+    input:
+        annotation_bed=list(config.get("annotate_cnv", {}).values()),
+        fai=config.get("reference", {}).get("fai", ""),
+        germline_vcf="snv_indels/bcbio_variation_recall_ensemble/{sample}_{type}.ensembled.vep_annotated.filter.germline.vcf",
+        json=expand(
+            "cnv_sv/cnv_html_report/{{sample}}_{{type}}.{caller}.json",
+            caller=config.get("svdb_merge", {}).get("cnv_callers", []),
+        ),
+        svdb_vcfs=expand(
+            "cnv_sv/svdb_query/{{sample}}_{{type}}.svdb_query.annotate_cnv.{tag}.vcf",
+            tag=list(config.get("annotate_cnv", {}).keys()),
+        ),
+    output:
+        json=temp("cnv_sv/cnv_html_report/{sample}_{type}.merged.json"),
+    params:
+        skip_chromosomes=config.get("reference", {}).get("skip_chrs", None),
+    log:
+        "cnv_sv/cnv_html_report/{sample}_{type}.merge_json.log",
+    benchmark:
+        repeat(
+            "cnv_sv/cnv_html_report/{sample}_{type}.merge_json.benchmark.tsv",
+            config.get("merge_json", {}).get("benchmark_repeats", 1),
+        )
+    threads: config.get("merge_json", {}).get("threads", config["default_resources"]["threads"])
+    resources:
+        mem_mb=config.get("merge_json", {}).get("mem_mb", config["default_resources"]["mem_mb"]),
+        mem_per_cpu=config.get("merge_json", {}).get("mem_per_cpu", config["default_resources"]["mem_per_cpu"]),
+        partition=config.get("merge_json", {}).get("partition", config["default_resources"]["partition"]),
+        threads=config.get("merge_json", {}).get("threads", config["default_resources"]["threads"]),
+        time=config.get("merge_json", {}).get("time", config["default_resources"]["time"]),
+    container:
+        config.get("merge_json", {}).get("container", config["default_container"])
+    conda:
+        "../envs/merge_json.yaml"
+    message:
+        "{rule}: Merge JSON representations from multiple callers: {output.json}"
+    script:
+        "../scripts/merge_json.py"
+
+
 rule cnv_html_report:
     input:
-        json="cnv_sv/cnv_html_report/{sample}_{type}.cnv.json",
+        json="cnv_sv/cnv_html_report/{sample}_{type}.merged.json",
         template=config.get("cnv_html_report", {}).get("template", ""),
     output:
         html=temp("cnv_sv/cnv_html_report/{sample}_{type}.cnv.html"),
