+++ workflow/scripts/merge_json.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 import cyvcf2
 import json
 
@@ -28,26 +29,51 @@ def get_vaf(vcf_filename, skip=None):
         yield variant.CHROM, variant.POS, variant.INFO.get("AF", None)
 
 
-def merge_cnv_dicts(dicts, vaf, annotations, chromosomes):
+def get_svdb_cnvs(vcf_filename, skip=None):
+    cnvs = defaultdict(list)
+    vcf = cyvcf2.VCF(vcf_filename)
+    for variant in vcf:
+        if not skip is None and variant.CHROM in skip:
+            continue
+        caller = variant.INFO.get("CALLER")
+        genes = variant.INFO.get("Genes")
+        if genes is None:
+            continue
+        cnvs[caller].append(
+            dict(
+                chromosome=variant.CHROM,
+                genes=sorted(genes.split(",")),
+                start=variant.POS,
+                length=variant.INFO.get("SVLEN"),
+                type=variant.INFO.get("SVTYPE"),
+                copy_number=variant.INFO.get("CORR_CN"),
+            )
+        )
+    return cnvs
+
+
+def merge_cnv_dicts(dicts, vaf, annotations, chromosomes, svdb_cnvs):
     callers = list(map(lambda x: x["caller"], dicts))
     cnvs = {}
     for chrom, chrom_length in chromosomes:
         cnvs[chrom] = dict(
             chromosome=chrom,
+            label=chrom,
             length=chrom_length,
             vaf=[],
             annotations=[],
-            callers={c: dict(name=c, ratios=[], segments=[]) for c in callers},
+            callers={c: dict(name=c, ratios=[], segments=[], cnvs=[]) for c in callers},
         )
 
     for a in annotations:
-        cnvs[a[0]]["annotations"].append(
-            dict(
-                start=a[1],
-                end=a[2],
-                name=a[3],
+        for item in a:
+            cnvs[item[0]]["annotations"].append(
+                dict(
+                    start=item[1],
+                    end=item[2],
+                    name=item[3],
+                )
             )
-        )
 
     for v in vaf:
         cnvs[v[0]]["vaf"].append(
@@ -56,6 +82,19 @@ def merge_cnv_dicts(dicts, vaf, annotations, chromosomes):
                 vaf=v[2],
             )
         )
+
+    for d in svdb_cnvs:
+        for caller, cnv_list in d.items():
+            for c in cnv_list:
+                cnvs[c["chromosome"]]["callers"][caller]["cnvs"].append(
+                    dict(
+                        genes=c["genes"],
+                        start=c["start"],
+                        length=c["length"],
+                        type=c["type"],
+                        cn=c["copy_number"],
+                    )
+                )
 
     for d in dicts:
         for r in d["ratios"]:
@@ -82,10 +121,11 @@ def merge_cnv_dicts(dicts, vaf, annotations, chromosomes):
 
 
 def main():
-    annotation_bed = snakemake.input["annotation_bed"]
+    annotation_beds = snakemake.input["annotation_bed"]
     fasta_index_file = snakemake.input["fai"]
     germline_vcf = snakemake.input["germline_vcf"]
     json_files = snakemake.input["json"]
+    svdb_vcf_files = snakemake.input["svdb_vcfs"]
 
     output_file = snakemake.output["json"]
 
@@ -99,10 +139,17 @@ def main():
     fai = parse_fai(fasta_index_file, skip_chromosomes)
     vaf = get_vaf(germline_vcf)
     annotations = []
-    if len(annotation_bed) > 0:
-        annotations = parse_annotation_bed(annotation_bed, skip_chromosomes)
+    for filename in annotation_beds:
+        annotations.append(parse_annotation_bed(filename, skip_chromosomes))
 
-    cnvs = merge_cnv_dicts(cnv_dicts, vaf, annotations, fai)
+    svdb_cnvs = []
+    for filename in svdb_vcf_files:
+        # Parse VCF file and get annotated, called CNVs. If there are none,
+        # create an empty list for each caller. In this case, no table will
+        # be presented in the final report.
+        svdb_cnvs.append(get_svdb_cnvs(filename, skip_chromosomes))
+
+    cnvs = merge_cnv_dicts(cnv_dicts, vaf, annotations, fai, svdb_cnvs)
 
     with open(output_file, "w") as f:
         print(json.dumps(cnvs), file=f)
